[Refactor] Removed as much internals as possible out of public API
