+++ include/occa/core/device.hpp
@@ -12,7 +12,7 @@
 #include <occa/types.hpp>
 
 // Unfortunately we need to expose this in include
-#include <occa/internal/utils/gc.hpp>
+#include <occa/utils/gc.hpp>
 
 namespace occa {
   class modeKernel_t; class kernel;
