+++ include/occa/core/memory.hpp
@@ -8,7 +8,7 @@
 #include <occa/types.hpp>
 
 // Unfortunately we need to expose this in include
-#include <occa/internal/utils/gc.hpp>
+#include <occa/utils/gc.hpp>
 
 namespace occa {
   class modeMemory_t; class memory;
