+++ include/occa/core/stream.hpp
@@ -7,7 +7,7 @@
 #include <occa/types.hpp>
 
 // Unfortunately we need to expose this in include
-#include <occa/internal/utils/gc.hpp>
+#include <occa/utils/gc.hpp>
 
 namespace occa {
   class modeStream_t; class stream;
