+++ src/occa/internal/utils/gc.cpp
@@ -42,9 +42,5 @@ namespace occa {
       }
       leftRingEntry = rightRingEntry = this;
     }
-
-    bool ringEntry_t::isAlone() const {
-      return (leftRingEntry == this);
-    }
   }
 }
