+++ src/occa/internal/utils/string.hpp
@@ -1,5 +1,5 @@
-#ifndef OCCA_INTERNAL_UTILS_STRING_HEADER
-#define OCCA_INTERNAL_UTILS_STRING_HEADER
+#ifndef OCCA_FAKEINTERNAL_UTILS_STRING_HEADER
+#define OCCA_FAKEINTERNAL_UTILS_STRING_HEADER
 
 #include <iostream>
 #include <iomanip>
